split out base conversion methods
